requirements.txt
Flask~=2.0.1
PyJWT~=2.1.0
bcrypt~=3.2.0
SQLAlchemy~=1.4.19
pytz~=2021.1